+++ packages/core/src/Accordion.svelte
@@ -1,7 +1,7 @@
 <script>
-	import { Text, Summary } from './items'
 	import { defaultFields } from './constants'
-	import ListItems from './ListItems.svelte'
+	import { Text, Summary } from './items'
+	import List from './List.svelte'
 
 	let className = ''
 	export { className as class }
@@ -10,23 +10,39 @@
 	export let using = {}
 	export let autoClose = false
 	export let value = null
-
-	let activeGroup = null
+	// let activeGroup = null
 
 	$: fields = { ...defaultFields, ...fields }
 	$: using = { default: Text, ...using }
+	// $: openGroupForSeleceted(items)
 
-	function handleToggle(event) {
+	// function openGroupForSeleceted(items) {
+	// 	if (value) {
+	// 		let matched = items.find(
+	// 			(x) => x[fields.data].findIndex((y) => y == value) > -1
+	// 		)
+	// 		if (matched) {
+	// 			toggle(matched)
+	// 			matched[fields.isOpen] = true
+	// 			// if (typeof window !== 'undefined') alert(matched)
+	// 		}
+	// 	}
+	// }
+	function toggle(item) {
 		if (autoClose) {
-			if (
-				activeGroup &&
-				activeGroup !== event.detail &&
-				activeGroup[fields.isOpen]
-			) {
-				activeGroup[fields.isOpen] = false
+			// this event is triggered before the state changes
+			if (!item[fields.isOpen]) {
+				items.map((x) => {
+					if (x !== item && x[fields.isOpen]) {
+						x[fields.isOpen] = false
+					}
+				})
 			}
-			activeGroup = event.detail
+			// if (activeGroup && activeGroup !== item && activeGroup[fields.isOpen]) {
+			// 	activeGroup[fields.isOpen] = false
+			// }
 		}
+		// activeGroup = item
 	}
 </script>
 
@@ -35,10 +51,20 @@
 		{@const hasItems = item[fields.data] && item[fields.data].length > 0}
 		{@const itemFields = fields.fields ?? fields}
 
-		<details class="flex flex-col" class:expanded={item[fields.isOpen]}>
-			<Summary {fields} {using} bind:content={item} on:toggle={handleToggle} />
+		<details
+			class="flex flex-col"
+			class:is-expanded={item[fields.isOpen]}
+			bind:open={item[fields.isOpen]}
+		>
+			<Summary
+				{fields}
+				{using}
+				bind:content={item}
+				on:click={() => toggle(item)}
+			/>
+
 			{#if hasItems && item[fields.isOpen]}
-				<ListItems
+				<List
 					bind:items={item[fields.data]}
 					bind:value
 					fields={itemFields}