+++ packages/markdown/src/CodeSnippet.svelte
@@ -1,11 +1,13 @@
 <script>
-	import { highlight, languages } from 'prismjs'
+	import * as prism from 'prismjs'
 	// import 'prism-svelte'
 
 	export let code
 	export let language
+
+	$: parsed = prism.highlight(code, prism.languages[language], language)
 </script>
 
 <code class="language-{language} font-mono">
-	{@html highlight(code, languages[language], language)}
+	{@html parsed}
 </code>