+++ packages/markdown/src/index.js
@@ -1,3 +1,4 @@
 export { default as Markdown } from './Markdown.svelte'
 export { default as Fillable } from './Fillable.svelte'
 export { default as CodeSnippet } from './CodeSnippet.svelte'
+export { default as ComponentExamples } from './ComponentExamples.svelte'