+++ sites/rokkit/src/app.d.ts
@@ -1,6 +1,8 @@
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 // and what to do when importing types
+import { type Sample } from '$lib/types'
+
 interface ComponentData {
 	icon: string
 	category: string
@@ -9,13 +11,23 @@ interface ComponentData {
 	url: string
 	component: string
 }
+interface ComponentMenu {
+	name: string
+	id: string
+	data: ComponentData[]
+}
 
+interface Sections {
+	slug: string
+	title: string
+}
 declare namespace App {
 	// interface Error {}
 	// interface Locals {}
 	interface PageData {
-		menu: ComponentData[]
-		current?: ComponentData
+		sections: Sections[]
+		menu?: ComponentMenu[]
+		examples?: Sample[]
 	}
 	// interface Platform {}
 }