+++ sites/rokkit/src/routes/(content)/+layout.js
@@ -1,13 +1,9 @@
-import { components } from '$lib'
+import { components, categories } from '$lib'
 
 /** @type {import('./$types').LayoutLoad} */
 export function load() {
-	const menu = components.map((m) => ({
-		...m,
-		url: '/' + m.name.toLowerCase(),
-		component: 'link'
-	}))
 	return {
-		menu
+		components,
+		categories
 	}
 }